+++ CHANGELOG.md
@@ -95,6 +95,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   The rendering behavior remains unchanged, and the image will scale correctly
   within whatever bounds it is given.
+- `Widget::unmounted` is now invoked for all widgets in the hierarchy.
+  Previously, only the parent widget was having its unmounted event invoked.
 
 ### Changed